Remove TensorFlow support
