+++ app/ui/src/components/Bot/Playground/Menu.tsx
@@ -8,7 +8,7 @@ export const PlaygroundMenu = () => {
   return (
     <div
       id="menu"
-      className={`bg-white z-[999] border fixed md:inset-y-0 md:flex md:w-[260px] md:flex-col transition-transform  max-md:w-3/4  ${
+      className={`bg-white z-[998] border fixed md:inset-y-0 md:flex md:w-[260px] md:flex-col transition-transform  max-md:w-3/4  ${
         hideMenu ? "translate-x-[-100%]" : "translate-x-[0%]"
       }`}
     >
