+++ app/ui/src/utils/embeddings.ts
@@ -1,6 +1,5 @@
 export const availableEmbeddingTypes = [
   { value: "openai", label: "OpenAI" },
-  { value: "tensorflow", label: "Tensorflow (cpu)" },
   { value: "cohere", label: "Cohere" },
   { value: "huggingface-api", label: "HuggingFace (Inference)" },
   {
