+++ server/package.json
@@ -39,10 +39,6 @@
     "@prisma/client": "5",
     "@slack/bolt": "^3.13.2",
     "@supabase/supabase-js": "^2.24.0",
-    "@tensorflow-models/universal-sentence-encoder": "^1.3.3",
-    "@tensorflow/tfjs-backend-cpu": "^4.7.0",
-    "@tensorflow/tfjs-converter": "^4.7.0",
-    "@tensorflow/tfjs-core": "^4.7.0",
     "@types/html-to-text": "^9.0.4",
     "@waylaidwanderer/fastify-sse-v2": "^3.1.0",
     "@xenova/transformers": "^2.7.0",
