+++ server/prisma/seed.ts
@@ -47,14 +47,14 @@ const MODELS: {
     config: "{}",
   },
   {
-    model_id:"accounts/fireworks/models/mixtral-8x7b-fw-chat",
+    model_id: "accounts/fireworks/models/mixtral-8x7b-fw-chat",
     name: "Mixtral MoE 8x7B Chat (Fireworks)",
     model_type: "chat",
     stream_available: true,
     model_provider: "Fireworks",
     local_model: false,
     config: "{}",
-  }
+  },
 ];
 
 const newModels = async () => {
@@ -70,8 +70,20 @@ const newModels = async () => {
   }
 };
 
+const removeTensorflowSupport = async () => {
+  await prisma.bot.updateMany({
+    where: {
+      embedding: "tensorflow",
+    },
+    data: {
+      embedding: "transformer",
+    },
+  });
+};
+
 const main = async () => {
   await newModels();
+  await removeTensorflowSupport();
 };
 
 main()
