+++ server/src/routes/api/v1/bot/handlers/get.handler.ts
@@ -141,7 +141,6 @@ export const getCreateBotConfigHandler = async (
 
   const embeddingModel = [
     { value: "openai", label: "text-embedding-ada-002" },
-    { value: "tensorflow", label: "tensorflow (cpu)" },
     { value: "cohere", label: "Cohere" },
     {
       value: "transformer",
@@ -198,7 +197,6 @@ export const getBotByIdSettingsHandler = async (
 
   const embeddingModel = [
     { value: "openai", label: "text-embedding-ada-002" },
-    { value: "tensorflow", label: "tensorflow (cpu)" },
     { value: "cohere", label: "Cohere" },
     {
       value: "transformer",
